feat: fix root redirect, dynamic intelligence routing, and discover architecture
- / now redirects to /app/login (middleware handles authed → /app/feed)
- /app/intelligence/[slug] is now a dynamic route — accepts any topic slug,
  queries Supabase by slug, returns 404 on miss
- /app/intelligence redirects to /app/discover (sidebar nav still highlights
  Intelligence when on any /app/intelligence/* path)
- Discover search navigates to /app/intelligence/[slug] instead of /app/feed
- Feed "Explore Topic" CTA updated to /app/intelligence/[slug]
- Intelligence page adds Discover ← breadcrumb back nav
- All hardcoded "canada-immigration" slug references replaced with dynamic param
